update moments function
